(PC-36858)[PRO] refactor: factorize duplicate form event publication
