+++ pro/src/components/IndividualOfferLayout/components/OfferPublicationEdition/OfferPublicationEditionForm/OfferPublicationEditionForm.module.scss
@@ -15,20 +15,6 @@
   gap: var(--size-spacing-xl);
 }
 
-.date-picker {
-  width: rem.torem(180px);
-}
-
-.time-picker {
-  width: rem.torem(110px);
-}
-
-.inputs-row {
-  display: flex;
-  gap: var(--size-spacing-xl);
-}
-
-.group,
 .toggle {
   margin-bottom: var(--size-spacing-xxl);
 }
