+++ pro/src/components/PublicationAndBookingFields/PublicationAndBookingFields.module.css
@@ -1,0 +1,3 @@
+.group {
+  margin-bottom: var(--size-spacing-xxl);
+}
