+++ pro/src/pages/IndividualOffer/IndividualOfferSummary/components/EventPublicationForm/EventPublicationForm.module.scss
@@ -4,18 +4,6 @@
   margin-top: var(--size-spacing-xxl);
 }
 
-.publish-later {
-  margin-top: var(--size-spacing-l);
-}
-
-.date-picker {
-  width: rem.torem(180px);
-}
-
-.time-picker {
-  width: rem.torem(110px);
-}
-
 .title-container {
   display: flex;
   align-items: center;
