+++ pro/src/ui-kit/form/DateAndTimePicker/DateAndTimePicker.module.scss
@@ -1,0 +1,14 @@
+@use "styles/mixins/_rem.scss" as rem;
+
+.inputs-row {
+  display: flex;
+  gap: var(--size-spacing-xl);
+}
+
+.date-picker {
+  width: rem.torem(180px);
+}
+
+.time-picker {
+  width: rem.torem(110px);
+}
